Several flights, some problems to resolve, and several deconnections.

+++ tools.cpp
@@ -6,43 +6,43 @@ void tools_int16tToBuffer(int16_t value, uint8_t* buffer, bool incBuffer){
 	if(incBuffer) buffer += length;
 }
 
-void tools_uint16tToBuffer(int16_t value, uint8_t* buffer, bool incBuffer){
+void tools_uint16tToBuffer(uint16_t value, uint8_t* buffer, bool incBuffer){
 	const uint8_t length = 2;
 	memcpy(buffer, &value, length);
 	if(incBuffer) buffer += length;
 }
 
-void tools_int32tToBuffer(int16_t value, uint8_t* buffer, bool incBuffer){
+void tools_int32tToBuffer(int32_t value, uint8_t* buffer, bool incBuffer){
 	const uint8_t length = 4;
 	memcpy(buffer, &value, length);
 	if(incBuffer) buffer += length;
 }
 
-void tools_uint32tToBuffer(int16_t value, uint8_t* buffer, bool incBuffer){
+void tools_uint32tToBuffer(uint32_t value, uint8_t* buffer, bool incBuffer){
 	const uint8_t length = 4;
 	memcpy(buffer, &value, length);
 	if(incBuffer) buffer += length;
 }
 
-void tools_int64tToBuffer(int16_t value, uint8_t* buffer, bool incBuffer){
+void tools_int64tToBuffer(int64_t value, uint8_t* buffer, bool incBuffer){
 	const uint8_t length = 8;
 	memcpy(buffer, &value, length);
 	if(incBuffer) buffer += length;
 }
 
-void tools_uint64tToBuffer(int16_t value, uint8_t* buffer, bool incBuffer){
+void tools_uint64tToBuffer(uint64_t value, uint8_t* buffer, bool incBuffer){
 	const uint8_t length = 8;
 	memcpy(buffer, &value, length);
 	if(incBuffer) buffer += length;
 }
 
-void tools_floatToBuffer(int16_t value, uint8_t* buffer, bool incBuffer){
+void tools_floatToBuffer(float value, uint8_t* buffer, bool incBuffer){
 	const uint8_t length = 4;
 	memcpy(buffer, &value, length);
 	if(incBuffer) buffer += length;
 }
 
-void tools_doubleToBuffer(int16_t value, uint8_t* buffer, bool incBuffer){
+void tools_doubleToBuffer(double value, uint8_t* buffer, bool incBuffer){
 	const uint8_t length = 8;
 	memcpy(buffer, &value, length);
 	if(incBuffer) buffer += length;
