+++ tools.h
@@ -6,16 +6,16 @@
 #include "parrot_esp.h"
 
 void tools_int16tToBuffer(int16_t value, uint8_t* buffer, bool incBuffer = false);
-void tools_uint16tToBuffer(int16_t value, uint8_t* buffer, bool incBuffer = false);
+void tools_uint16tToBuffer(uint16_t value, uint8_t* buffer, bool incBuffer = false);
 
-void tools_int32tToBuffer(int16_t value, uint8_t* buffer, bool incBuffer = false);
-void tools_uint32tToBuffer(int16_t value, uint8_t* buffer, bool incBuffer = false);
+void tools_int32tToBuffer(int32_t value, uint8_t* buffer, bool incBuffer = false);
+void tools_uint32tToBuffer(uint32_t value, uint8_t* buffer, bool incBuffer = false);
 
-void tools_int64tToBuffer(int16_t value, uint8_t* buffer, bool incBuffer = false);
-void tools_uint64tToBuffer(int16_t value, uint8_t* buffer, bool incBuffer = false);
+void tools_int64tToBuffer(int64_t value, uint8_t* buffer, bool incBuffer = false);
+void tools_uint64tToBuffer(uint64_t value, uint8_t* buffer, bool incBuffer = false);
 
-void tools_floatToBuffer(int16_t value, uint8_t* buffer, bool incBuffer = false);
-void tools_doubleToBuffer(int16_t value, uint8_t* buffer, bool incBuffer = false);
+void tools_floatToBuffer(float value, uint8_t* buffer, bool incBuffer = false);
+void tools_doubleToBuffer(double value, uint8_t* buffer, bool incBuffer = false);
 
 
 int16_t tools_bufferToInt16t(uint8_t* buffer, bool incBuffer = false);
